Seattle Air BnB Initial Investment strategy
An overview project using datasets found in the public domain on rentals in Seattle in the form of an Excel spreadsheet Data was used to create 5 sheets and combined into a workbook with an exported PowerPoint file to view the final result and showcase basic skillset. The original Excel sheet is to large to upload to git. Sheets in the workbook:
1. ZipCode and AVG price(based on calendar used to create Price by Zip Code bar graph
2. price per zip code color-coded map sorted by avg price and zip code
3. Line graph showing revenue for the year in a weekly format
4. bar graph using bedrooms and avg price to show avg price based on bedroom count
5. A simple Distinct count showing the number of listings available sorted by bedroom count